% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JonsSSUtils.R
\docType{data}
\name{example_ss_files}
\alias{example_ss_files}
\title{Example Stock Synthesis files as an r4ss object}
\format{
A multi-level list. The top four levels correspond to the starter,
    data, control and forecast files.
}
\usage{
data("example_ss_files")
}
\description{
The 2025 NT Black Jewfish Stock Assessment SS files that can be used as example
data.
}
\keyword{datasets}
