% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JonsSSUtils.R
\docType{package}
\name{JonsSSUtils}
\alias{JonsSSUtils-package}
\alias{JonsSSUtils}
\title{JonsSSUtils}
\description{
A package to estimate fish length-at-age models using MCMC analysis with 'rstan' models.

Standardised functions to simply Jonathan Smart's Stock Synthesis workflow and
    provide this to collaborators.
}
\author{
\strong{Maintainer}: Jonathan Smart \email{jonsmartphd@gmail.com}

}
