% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Profile_SS_Parameters}
\alias{Profile_SS_Parameters}
\title{Profile_SS_Parameters}
\usage{
Profile_SS_Parameters(
  Origin = getwd(),
  SS_dr = NULL,
  string = NULL,
  vector = NULL,
  Piner = TRUE,
  Comparisons = TRUE,
  profile.label = NULL,
  legendlabels = NULL,
  SS_loc = "/Users/jonathansmart/Documents/SS/",
  ss_exe = "ss3_osx_arm64",
  print_r4SS_to_screen = FALSE,
  copy_to_origin = TRUE
)
}
\arguments{
\item{Origin}{The working directory on a network or hard drive where the
Stock Synthesis assessment is being performed. The SS files will be copied
from here to `SS_dr` where the model is run. If `copy_to_origin == TRUE`
(default) then the SS model files are copied back to this directory.}

\item{SS_dr}{The location on the hard drive where the SS model will be run locally.
make sure there is no white space in this filepath, especially if using a
non Windows computer.}

\item{string}{A string that determines which parameter within the model is
profiled.}

\item{vector}{The vector of parameter values to be profiled}

\item{Piner}{Should Piner plots be produced}

\item{Comparisons}{Should model comparison plots be produced?}

\item{profile.label}{Label for the plot titles (typically the parameter being profiled)}

\item{legendlabels}{Labels used for the plot legends}

\item{SS_loc}{The location of the SS executable file.}

\item{ss_exe}{The name of the SS executable. The default is 'ss3_osx_arm64'
which corresponds to my own operating system.}

\item{print_r4SS_to_screen}{How much information from `r4ss` would you like
vommitted back at you?}

\item{copy_to_origin}{Do you want the SS files (and `r4ss` plots if requested)
copied back to the `Origin`? This can take a while if working remotely with
an agency network.}
}
\value{
Plots saved to created directories of the parameter profiling. No
    R objects are returned.
}
\description{
This is a wrapper function around `r4ss::profile` which, again,
    runs stock synthesis from the designated location on your computer and stores
    the outputs in the directory that the stock assessment is being developed.
    The profile plots for the chosen parameter are automatically saved in new
    locations on the computer.
}
