% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Run_Stock_Synthesis}
\alias{Run_Stock_Synthesis}
\title{Run_Stock_Synthesis}
\usage{
Run_Stock_Synthesis(
  Origin = getwd(),
  SS_dr = NULL,
  SS_loc = "/Users/jonathansmart/Documents/SS/",
  do_hess = TRUE,
  do_forecast = FALSE,
  plots = TRUE,
  ss_exe = "ss3_osx_arm64",
  print_r4SS_to_screen = FALSE,
  print_updates_to_screen = TRUE,
  Ask = TRUE,
  copy_to_origin = TRUE,
  ...
)
}
\arguments{
\item{Origin}{The working directory on a network or hard drive where the
Stock Synthesis assessment is being performed. The SS files will be copied
from here to `SS_dr` where the model is run. If `copy_to_origin == TRUE`
(default) then the SS model files are copied back to this directory.}

\item{SS_dr}{The location on the hard drive where the SS model will be run locally.
make sure there is no white space in this filepath, especially if using a
non Windows computer.}

\item{SS_loc}{The location of the SS executable file.}

\item{do_hess}{Should the hessian be calculated}

\item{do_forecast}{Should forecasts be included in the `r4ss` outputs?}

\item{plots}{Do you want to produce the `r4ss` outputs?}

\item{ss_exe}{The name of the SS executable. The default is 'ss3_osx_arm64'
which corresponds to my own operating system.}

\item{print_r4SS_to_screen}{How much information from `r4ss` would you like
vommitted back at you?}

\item{print_updates_to_screen}{How much information from the SS model run would
you like vommitted back at you.}

\item{Ask}{Set to `TRUE` (default) so that the function prompts you to check the
`SS_dr` is correct. All existing files are erased from here at the beginning
of the function call so this is way of avoiding accidentally wiping files from
an incorrect file path.}

\item{copy_to_origin}{Do you want the SS files (and `r4ss` plots if requested)
copied back to the `Origin`? This can take a while if working remotely with
an agency network.}

\item{...}{Function calls to be passed to `r4ss`}
}
\value{
An SS model run saved to the relevant directories as well as `r4ss`
    plots if requested. No R objects are returned.
}
\description{
The Stock Synthesis software is getting harder to run on modern
    IT setups given fears around malicious software. Often stock assessments are
    best developed in a shared space such as a network or repository. However,
    often the SS executable can't be run from these locations and a specific
    directory on a computer is needed so that the file can be marked as safe.
    Having worked in many agencies, each with different IT permissions, the best
    way I've found to handle this is to develop the SS files in whatever location
    is most convenient (such as a shared network directory), copy those files to
    the safe folder where SS can be run, and then copy the SS files back to the
    original location. This is a bit circular but it works, and this code makes that
    process simple. It will also produce the `r4ss` plots in one hit and transfer
    these too.
}
